Creating two tags at the same time now works
Concurrency! Before this it was possible for two tags to be created at
the same time due to the "try to get it, if it doesn't exist then create
it" style. Now it's a "try to create it; if the database yells at you
then it exists so get it." thing. Also, there is a constraint on the db
now.